filters/aesthetic/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
braceexpand==0.1.7
certifi==2025.1.31
charset-normalizer==3.4.1
clip==0.2.0
cmake==4.0.0
contourpy==1.3.1
cycler==0.12.1
datasets==2.10.1
dill==0.3.6
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
ftfy==6.3.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
lit==18.1.8
MarkupSafe==3.0.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.14
networkx==3.4.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-clip==1.0.1
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==1.5.3
Pillow==9.4.0
propcache==0.3.1
pyarrow==19.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
six==1.17.0
sympy==1.13.3
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.13.1
urllib3==2.3.0
wcwidth==0.2.13
webdataset==0.2.48
xxhash==3.5.0
yarl==1.19.0
